+++ src/util/searchProviders/searchDuckDuckGo.ts
@@ -1,11 +1,9 @@
 import { chromium, Browser, Page, LaunchOptions } from 'playwright';
 import logger from '../logger';
 import { SearchResult } from "../scourFormat";
-import { determineHeadlessShellPath } from '../determineHeadlessShellPath';
+import { determineBrowserPath } from '../determineBrowserPath';
 import { stealthifyPlaywright } from '../stealthifyPlaywright';
 import { readScourFile, writeScourFile } from '../scourFileUtils';
-
-const SHOW_BROWSER = true;
 
 function extractBareUrl(duckDuckGoUrl: string): string {
     if (duckDuckGoUrl.substring(0, 2) === '//') {
@@ -28,15 +26,12 @@ export async function searchDuckDuckGo(query: string, numResultsPerQuery: number
     logger.info("Launching browser...");
     let executablePath = null;
 
-    // If the browser is to be shown, the headless shell path cannot be used.
-    if (!SHOW_BROWSER) {
-        executablePath = determineHeadlessShellPath();
-    }
+    executablePath = determineBrowserPath();
 
     const scourFile = readScourFile(outputFilePath);
 
-    if (!SHOW_BROWSER && !executablePath) {
-        logger.error('Failed to find headless shell path.');
+    if (!executablePath) {
+        logger.error('Failed to find Chromium path.');
         scourFile.statusMessage = 'Failed to find path to Chromium. Please make sure Chromium is installed correctly via Playwright.';
         writeScourFile(outputFilePath, scourFile);
         return [];
@@ -44,10 +39,10 @@ export async function searchDuckDuckGo(query: string, numResultsPerQuery: number
 
     try {
         const options: LaunchOptions = {
-            headless: !SHOW_BROWSER,
+            headless: false,
         };
 
-        if (!SHOW_BROWSER && executablePath) {
+        if (executablePath) {
             options.executablePath = executablePath;
         }
